Extract RecordTransformBuffer interface
By extracting the `RecordTransformBuffer` from the class that implements it, now called `SimpleRecordTransformBuffer`, we do the following:

* Unlock experimentation with other implementations of `RecordTransformBuffer`.
* Remove the circular dependency between `sync-record-cache` and `record-transform-buffer`. This cycle, while still "legal", highlighted an issue in esbuild.
* Require that an _instance_ of a `RecordTransformBuffer` be passed into record caches on construction, which simplifies the settings.

Note: To keep behavior consistent, all the standard derived cache classes now override `_getTransformBuffer` to construct a `SimpleRecordTransformBuffer` if a `transformBuffer` hasn't been provided in settings.
