+++ packages/@orbit/memory/src/memory-cache.ts
@@ -11,6 +11,8 @@ import {
   RecordCacheTransformOptions,
   RecordCacheUpdateDetails,
   RecordRelationshipIdentity,
+  RecordTransformBuffer,
+  SimpleRecordTransformBuffer,
   SyncRecordCache,
   SyncRecordCacheSettings
 } from '@orbit/record-cache';
@@ -255,6 +257,22 @@ export class MemoryCache<
   /////////////////////////////////////////////////////////////////////////////
   // Protected methods
   /////////////////////////////////////////////////////////////////////////////
+
+  /**
+   * Override `_getTransformBuffer` on base `SyncRecordCache` to provide a
+   * `transformBuffer` if a custom one hasn't been provided via the constructor
+   * setting.
+   */
+  protected _getTransformBuffer(): RecordTransformBuffer {
+    if (this._transformBuffer === undefined) {
+      const { schema, keyMap } = this;
+      this._transformBuffer = new SimpleRecordTransformBuffer({
+        schema,
+        keyMap
+      });
+    }
+    return this._transformBuffer;
+  }
 
   protected _resetInverseRelationships(
     base?: MemoryCache<QO, TO, QB, TB, QRD, TRD>
