+++ packages/@orbit/record-cache/src/index.ts
@@ -2,6 +2,7 @@ export * from './response';
 export * from './record-accessor';
 export * from './record-cache';
 export * from './record-transform-buffer';
+export * from './simple-record-transform-buffer';
 
 export * from './async-record-cache';
 export * from './async-operation-processor';
